#515 move fetch key table data to main page
